+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * `PySet::new` and `PyFrozenSet::new` now return `PyResult<&Py[Frozen]Set>`; exceptions are raised if
    the items are not hashable.
  * Fixed building using `venv` on Windows.
+ * `PyTuple::new` now returns `&PyTuple` instead of `Py<PyTuple>`.
 
 ## [0.6.0] - 2018-03-28